feat: Add Chatbot router use chatbot class
