+++ model-server/main.py
@@ -1,13 +1,13 @@
 from fastapi import FastAPI, Request
 import time
 import uvicorn
-from routers.model import predict
+from routers import chat
 
 
 app = FastAPI()
 
 app.include_router(
-    router=predict.router,
+    router=chat.router,
     prefix="/predict",
     tags=["predict"],
 )
@@ -28,4 +28,4 @@ def hello_world():
 
 
 if __name__ == "__main__":
-    uvicorn.run("main:app", host="0.0.0.0", port=8888, reload=True)
+    uvicorn.run("main:app", host="0.0.0.0", port=15421, reload=True)
