+++ model/src/models/cos_sim_chatbot.py
@@ -154,13 +154,13 @@ if __name__ == "__main__":
 
     root = pyrootutils.setup_root(__file__, pythonpath=True)
 
-    cfg = omegaconf.OmegaConf.create({"paths": {"data_dir": "null"},
-                                      "models": "null"})
-    cfg.paths.data_dir = "/home/sj/Project/2022-1-CECD3-EverySon-2/model/data/"
-    cfg.model = omegaconf.OmegaConf.load(root / "configs" / "model" / "cos_sim_chatbot.yaml")
-    model = hydra.utils.instantiate(cfg.model)
-    model = CosSimChatbot.load_from_checkpoint("logs/train/runs/2022-10-29_00-57-07/checkpoints/last.ckpt")
-    tokenizer = AutoTokenizer.from_pretrained(cfg.model.pretrained_model_name_or_path)
+    # cfg = omegaconf.OmegaConf.create({"paths": {"data_dir": "null", "output_dir": "logs/train/runs/2022-12-06_10-48-36/chatbot_db"},
+    #                                   "model": "null"})
+    # cfg.paths.data_dir = "/home/ubuntu/hanbin/2022-1-CECD3-EverySon-2/model/data/"
+    cfg = omegaconf.OmegaConf.load(root / "configs" / "model" / "cos_sim_chatbot.yaml")
+    model = hydra.utils.instantiate(cfg)
+    model = CosSimChatbot.load_from_checkpoint("logs/train/runs/2022-12-06_10-48-36/checkpoints/last.ckpt")
+    tokenizer = AutoTokenizer.from_pretrained(cfg.pretrained_model_name_or_path)
     
     text = "아 기분이 안좋네..."
     label = "슬픔"
@@ -174,8 +174,10 @@ if __name__ == "__main__":
         add_special_tokens=True,
         max_length=256,)
 
-    y = model.predict_step(tokenized_sentences, batch_idx=0)
-    print(y)
+    print(tokenized_sentences)
+    print(model.answer_dict)
+    # y = model.predict_step(tokenized_sentences, batch_idx=0)
+    # print(y)
 
-    model2 = CosSimChatbot.load_from_checkpoint("logs/train/runs/2022-10-29_00-57-07/checkpoints/last.ckpt")
-    print(model2.key)
+    # model2 = CosSimChatbot.load_from_checkpoint("logs/train/runs/2022-12-06_10-48-36/checkpoints/last.ckpt")
+    # print(model2.key)
